updated stracture
