--- ergasiaprog/questionnaire.html
@@ -1,95 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Questionnaire</title>
-    <link rel="stylesheet" href="style.css"> <!-- Add this line to link the CSS file -->
-</head>
-
-<h2 class="questionnaire-title">Τίτλος Ερωτηματολογίου</h2>
-<div class="questionnaire-container"></div>
-
-<body>
-    <div class="section1">
-        <div class="navbar">
-            <div class="logo">
-                <h1>Questionnaire</h1>
-            </div>
-        </div>
-        <div class="content">
-            <h1>Welcome to the Questionnaire</h1>
-            <p>Answer the following questions:</p>
-        </div>
-        <div class="card" id="card">
-            <div class="card-front" id="card-front">
-                <img src="image1.jpg" alt="Question 1 Image">
-                <h3>Question 1</h3>
-                <form>
-                    <label for="q1a">Option A</label>
-                    <input type="radio" id="q1a" name="question1" value="A" required><br>
-
-                    <label for="q1b">Option B</label>
-                    <input type="radio" id="q1b" name="question1" value="B" required><br>
-
-                    <label for="q1c">Option C</label>
-                    <input type="radio" id="q1c" name="question1" value="C" required><br>
-
-                    <img src="image2.jpg" alt="Question 2 Image">
-                    <h3>Question 2</h3>
-                    <label for="q2a">Option A</label>
-                    <input type="radio" id="q2a" name="question2" value="A" required><br>
-
-                    <label for="q2b">Option B</label>
-                    <input type="radio" id="q2b" name="question2" value="B" required><br>
-
-                    <label for="q2c">Option C</label>
-                    <input type="radio" id="q2c" name="question2" value="C" required><br>
-
-                    <img src="image3.jpg" alt="Question 3 Image">
-
-                    <h3>Question 3</h3>
-                    <label for="q3a">Option A</label>
-                    <input type="radio" id="q3a" name="question3" value="A" required><br>
-                    <label for="q3b">Option B</label>
-                    <input type="radio" id="q3b" name="question3" value="B" required><br>
-                    <label for="q3c">Option C</label>
-                    <input type="radio" id="q3c" name="question3" value="C" required><br>
-
-                    <img src="image4.jpg" alt="Question 4 Image">
-                    <h3>Question 4</h3>
-                    <label for="q4a">Option A</label>
-                    <input type="radio" id="q4a" name="question4" value="A" required><br>
-
-                    <label for="q4b">Option B</label>
-                    <input type="radio" id="q4b" name="question4" value="B" required><br>
-
-                    <label for="q4c">Option C</label>
-                    <input type="radio" id="q4c" name="question4" value="C" required><br>
-
-                    <img src="image5.jpg" alt="Question 5 Image">
-                    <h3>Question 5</h3>
-                    <label for="q5a">Option A</label>
-                    <input type="radio" id="q5a" name="question5" value="A" required><br>
-                    <label for="q5b">Option B</label>
-                    <input type="radio" id="q5b" name="question5" value="B" required><br>
-
-                    <label for="q5c">Option C</label>
-                    <input type="radio" id="q5c" name="question5" value="C" required><br>
-
-                    <!-- Repeat the above pattern for the remaining questions -->
-                </form>
-                <button type="button"><span></span>Submit</button>
-            </div>
-        </div>
-    </div>
-    
-    <link rel="stylesheet" href="style.css">
-
-    <script>
-        // Your JavaScript code here, if needed
-    </script>
-</body>
-
-</html>
