--- ergasiaprog/style.css
@@ -1,257 +1,0 @@
-* {
-    margin: 0;
-    padding: 0;
-    font-family: sans-serif;
-}
-
-body {
-    background-color:#009688; /* Set background color for the body */
-}
-
-.section1 {
-    width: 100%;
-    height: 100vh;
-    background-image: linear-gradient(rgba(0, 0, 0, 0.75), rgba(0, 0, 0, 0.75)), url(background.jpg);
-    background-size: cover;
-    background-position: left top; /* Μετακινήστε την εικόνα προς τα αριστερά πάνω */
-    display: flex;
-    align-items: center;
-    justify-content: flex-start; /* Μετακινήστε τα περιεχόμενα προς τα αριστερά */
-}
-
-
-.navbar {
-    width: 0%;
-    margin: auto;
-    padding: 0px 0;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-}
-
-.logo {
-    width: 220px;
-    color: rgb(251, 234, 214);
-    cursor: pointer;
-}
-
-.navbar ul li {
-    list-style: none;
-    display: inline-block;
-    margin: 100 120px;
-    position: relative;
-}
-
-.navbar ul li a {
-    text-decoration: none;
-    color: antiquewhite;
-    text-transform: capitalize;
-}
-
-.navbar ul li::after {
-    content: '';
-    height: 33px;
-    width: 30;
-    background: #009688;
-    position: absolute;
-    left: 30;
-    bottom: -35px;
-    transition: 30.3s;
-}
-
-.navbar ul li:hover:after {
-    width: 100%;
-}
-
-.content {
-    width: 100%;
-    position: absolute;
-    top: 20%;
-    transform: translateY(-50%);
-    text-align: center;
-    color: bisque;
-    align-items: center;
-    justify-content: center;
-}
-
-.card {
-    width: 100%;
-    height: 100%;
-    position: relative;
-    top: -13%;
-    transform: translateY(-05%);
-    text-align: center;
-    color: bisque;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    transition-duration: 11s;
-    perspective: 1000px;
-    transform-style: preserve-3d;
-}
-
-.card-front,
-.card-back {
-    background-color: rgba(100, 187, 126, 0.449);
-    width: 700px;
-    height: 500px;
-    box-shadow: 0 0 60px 40px rgba(0, 0, 0, 0.75);
-    border-radius: 20px;
-    text-align: center;
-    position: absolute;
-    box-sizing: border-box;
-    backface-visibility: hidden;
-}
-
-.card-back {
-    transform: rotateY(-200deg);
-}
-
-.card-front h3 {
-    font-weight: normal;
-    font-size: 24px;
-    text-align: center;
-    margin-top: 40px;
-}
-
-.card-front .register-link {
-    font-size: 14px;
-    color: bisque;
-    text-align: center;
-    margin-top: 35px;
-}
-
-.register-link p a {
-    color: bisque;
-    text-decoration: none;
-    font-weight: 600;
-}
-
-.register-link p a:hover {
-    text-decoration: underline;
-}
-
-.card-back h3 {
-    font-weight: normal;
-    font-size: 24px;
-    text-align: center;
-    margin-top: 30px;
-}
-
-.card-back .login-link {
-    font-size: 34px;
-    color: bisque;
-    text-align: center;
-    margin-top: 22px;
-}
-
-.login-link button,
-.register-link button {
-    width: 100px;
-    padding: 30px 0;
-    text-align: center;
-    margin: 20px 10px;
-    border-radius: 30px;
-    font-weight: normal;
-    border: 1px solid #8bd502;
-    background: transparent;
-    color: beige;
-    cursor: pointer;
-    position: relative;
-    overflow: hidden;
-}
-
-.input-box {
-    width: 60%;
-    background: transparent;
-    border: 3px solid hsl(0, 47%, 21%);
-    margin: 8px 0;
-    height: 60px;
-    border-radius: 45px;
-    padding: 20 30px;
-    box-sizing: border-box;
-    outline: none;
-    text-align: center;
-    color: rgb(234, 223, 208);
-}
-
-.input-box:hover {
-    border: 2px solid #009688;
-    box-shadow: 20 20 24px 22px #009688;
-}
-
-::placeholder {
-    color: rgb(175, 120, 53);
-    font-size: 12px;
-}
-
-button {
-    width: 400px;
-    padding: 15px 0;
-    text-align: center;
-    margin: 8px 10px;
-    border-radius: 30px;
-    font-weight: bold;
-    border: 2px solid #009688;
-    background: transparent;
-    color: rgb(113, 113, 78);
-    cursor: pointer;
-    position: relative;
-    overflow: hidden;
-}
-
-span {
-    background: #cbece8;
-    height: 100%;
-    width: 0;
-    border-radius: 30px;
-    position: absolute;
-    left: 0;
-    bottom: 0;
-    z-index: -1;
-    transition: 0.3s;
-}
-
-button:hover span {
-    width: 100%;
-}
-
-button:hover {
-    border: none;
-}
-
-.question {
-    margin-bottom: 20px;
-    color: #020e0d; /* Χρώμα κειμένου για τις ερωτήσεις */
-}
-
-label {
-    margin: 10px 0;
-    font-size: 16px;
-}
-
-input[type="radio"] {
-    margin: 0 5px;
-}
-
-.questionnaire-title {
-    margin-bottom: 60px;
-}
-form {
-    background-color: rgba(51, 20, 20, 0.9);
-    padding: 20px;
-    border-radius: 10px;
-    box-shadow: 0 0 20px rgba(69, 51, 51, 0.2);
-    margin-top: 20px;
-    max-width: 660px; /* Αυξήστε την τιμή για μεγαλύτερο πλάτος */
-    width: 100%; /* Καθιστά τη φόρμα να παίρνει το πλήρες πλάτος του container */
-    margin: 0 auto; /* Κεντράρει τη φόρμα οριζόντια */
-}
-
-.questionnaire-container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    height: 15vh; /* Use min-height instead of height to take at least 100% of the viewport height */\
-}
