--- webpage REV/QuizPage.html
@@ -1,34 +1,0 @@
-<html>
-    <head>
-        <meta charset="utf-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1">
-        <title>GrapeVin-Quiz</title>
-        <link rel="stylesheet" href="QuizStyle.css">
-        <body>
-            <div class="navbar">
-                <div class="logo">
-                    <h1><a href="index.html#section1" style="text-decoration: none; color: rgb(83, 184, 113);">GrapeVin</a></h1>
-                </div>
-                <ul>
-                    <li><a href="index.html#section1">Home</a></li>
-                    <li><a href="index.html#section2">About</a></li>
-                    <li><a href="index.html#section3">Contact</a></li>
-                </ul>
-            </div>
-            <div class="app">
-                <h1>Wine Quiz</h1>
-                <div class="quiz">
-                    <h2 id="question">Question 1</h2>
-                    <div id="answer-buttons">
-                        <button class="btn">Answer 1</button>
-                        <button class="btn">Answer 2</button>
-                        <button class="btn">Answer 3</button>
-                        <button class="btn">Answer 4</button>
-                    </div>
-                    <button id="next-btn">Next</button>
-                </div>
-            </div>
-            <script src="questionnaire.js"></script>
-        </body>
-    </head>
-</html>
