--- webpage REV/QuizStyle.css
@@ -1,103 +1,0 @@
-*{
-    margin: 0;
-    padding: 0;
-    font-family: sans-serif;
-    box-sizing: border-box;
-}
-.navbar{
-    width: 85%;
-    margin: auto;
-    padding: 20px 0;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-}
-.logo{
-    width: 100px;
-    color: rgb(83, 184, 113);
-    cursor: pointer;
-}
-.navbar ul li{
-    list-style: none;
-    display: inline-block;
-    margin: 0 20px;
-    position: relative;
-}
-.navbar ul li a{
-    text-decoration: none;
-    color: antiquewhite;
-    text-transform: capitalize;
-}
-.navbar ul li::after{
-    content: '';
-    height: 3px;
-    width: 0;
-    background: #009688;
-    position: absolute;
-    left: 0;
-    bottom: -5px;
-    transition: 0.3s;
-}
-.navbar ul li:hover:after{
-    width: 100%;
-}
-body{
-    background-image: linear-gradient(rgba(0,0,0,0.75),rgba(0,0,0,0.75)),url(winequiz.jpg);
-    background-size: cover;
-    background-position: center;
-}
-.app{
-    background: rgba(60, 147, 86, 0.409);
-    box-shadow: 0 0 40px 20px rgba(0,0,0,0.75);
-    width: 90%;
-    max-width: 600px;
-    margin: 100px auto 0;
-    border-radius: 20px;
-    padding: 30px;
-}
-.app h1{
-    font-size: 25px;
-    color: bisque;
-    font-weight: 600;
-    border-bottom: 1px solid #fff;
-    padding-bottom: 30px;
-    text-align: center;
-}
-.quiz{
-    padding: 20px 0;
-}
-.quiz h2{
-    font-size: 18px;
-    color: bisque;;
-    font-weight: 600;
-}
-.btn{
-    background: #fff;
-    color: #222;
-    font-weight: 500;
-    width: 100%;
-    border: 1px solid #222;
-    padding: 10px;
-    margin: 10px 0;
-    text-align: left;
-    border-radius: 10px;
-    cursor: pointer;
-    transition: 0.3s;
-}
-.btn:hover{
-    background: #d49339;
-    color: #fff;
-}
-#next-btn{
-    background: #5b9f98;
-    color: #fff;
-    font-weight: 500;
-    width: 150px;
-    border: 0;
-    padding: 10px;
-    margin: 20px auto 0;
-    border-radius: 30px;
-    cursor: pointer;
-    display:block;
-
-}
